refactor: Move wave related things to the wave module
Visibility changes to `pub`:
- Some fields in `WaveManager`
